pyproject.toml
[project]
name = "path"
version = "0.1.0"
description = "This is the corresponding .toml file required to create the project's uv environment."
requires-python = "==3.10"
dependencies = [
    'openpyxl==3.1.5',
    'pip-chill==1.0.3',
    'plotly==5.24.1',
    'python-dotenv==1.0.1',
    'scikit-learn==1.5.2',
    'statsmodels==0.14.4',
    'tensorflow[and-cuda]==2.15.1',
    'tensorflow-probability==0.23.0',
    'tqdm==4.66.5',
]
